+++ packages/broker/package.json
@@ -28,6 +28,7 @@
   "author": "Streamr Network AG <contact@streamr.com>",
   "license": "STREAMR NETWORK OPEN SOURCE LICENSE",
   "dependencies": {
+    "@ethersproject/hdnode": "^5.4.0",
     "@streamr/network-tracker": "1.0.0",
     "@streamr/slackbot": "^1.1.0",
     "@streamr/utils": "^0.0.1",
