+++ packages/protocol/src/utils/index.ts
@@ -2,7 +2,6 @@ import OrderingUtil from "./OrderingUtil"
 import StreamMessageValidator, { StreamMetadata } from "./StreamMessageValidator"
 import SigningUtil from "./SigningUtil"
 import { createTrackerRegistry, TrackerRegistry, SmartContractRecord } from "./TrackerRegistry"
-import { generateMnemonicFromAddress } from './NodeUtil'
 import { StreamID, toStreamID, StreamIDUtils } from "./StreamID"
 import { StreamPartID, toStreamPartID, StreamPartIDUtils, MAX_PARTITION_COUNT } from "./StreamPartID"
 import { KeyExchangeStreamIDUtils } from "./KeyExchangeStreamID"
@@ -16,7 +15,6 @@ export {
     SmartContractRecord,
     TrackerRegistry,
     createTrackerRegistry,
-    generateMnemonicFromAddress,
     toStreamID,
     toStreamPartID,
     StreamID,
